Backend/requirements.txt
Django==4.1.3
asgiref==3.5.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2022.6
sqlparse==0.4.3
django-cors-headers==4.0.0
openai
langchain
pandas
numpy

